Refactors the InfoBanner component: now it accepts a children element. Applies the Content component as a text container. Updates the Snapshots.
