--- src/components/InfoBanner/Highlight.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import { fontWeightBold } from '../../styles/designTokens';
-
-const Highlight = styled.span`
-  font-weight: ${fontWeightBold};
-`;
-
-export default Highlight;
